More reliable async state update API.
Don't pass the current state in directly: that encourages using that
state snapshot as the source of truth for state, when in fact the state
may change due to the update being async. Instead, the only way to
access state is through the update.state getter, which is guaranteed to
be up-to-date.